Created CurrencyPairTests
Moved existing tests relating to CurrencyPair to CurrencyPairTests
